requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==43.0.1
python-multipart==0.0.12
jinja2==3.1.4
python3-saml==1.16.0
httpx==0.27.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
aiohttp==3.10.5
aioping==0.4.0
apscheduler==3.10.4
dnspython==2.6.1
firebase-admin==6.4.0
openpyxl==3.1.5
reportlab==4.2.5
